dadobugbedrockores compatibility

+++ src/main/java/com/hollingsworth/arsnouveau/common/spell/effect/EffectBreak.java
@@ -67,16 +67,16 @@ public class EffectBreak extends AbstractEffect {
             if(spellStats.hasBuff(AugmentExtract.INSTANCE)) {
                 stack.enchant(Enchantments.SILK_TOUCH, 1);
                 state.getBlock().playerDestroy(world, getPlayer(shooter, (ServerLevel) world), pos1, world.getBlockState(pos1), world.getBlockEntity(pos1), stack);
-                destroyBlockSafely(world, pos1, false, shooter);
+                if(!state.is(BlockTagProvider.NO_BREAK_DROP))destroyBlockSafely(world, pos1, false, shooter);
             }else if(spellStats.hasBuff(AugmentFortune.INSTANCE)) {
                 int bonus = spellStats.getBuffCount(AugmentFortune.INSTANCE);
                 stack.enchant(Enchantments.BLOCK_FORTUNE, bonus);
                 state.getBlock().popExperience((ServerLevel) world, pos1, state.getExpDrop(world, pos1, bonus, 0));
                 state.getBlock().playerDestroy(world, getPlayer(shooter, (ServerLevel) world), pos1, world.getBlockState(pos1), world.getBlockEntity(pos1), stack);
-                destroyBlockSafely(world, pos1, false, shooter);
+                if(!state.is(BlockTagProvider.NO_BREAK_DROP))destroyBlockSafely(world, pos1, false, shooter);
             } else {
                 state.getBlock().playerDestroy(world, getPlayer(shooter, (ServerLevel) world), pos1, world.getBlockState(pos1), world.getBlockEntity(pos1), stack);
-                destroyBlockSafely(world, pos1, false, shooter);
+                if(!state.is(BlockTagProvider.NO_BREAK_DROP))destroyBlockSafely(world, pos1, false, shooter);
                 state.getBlock().popExperience((ServerLevel) world, pos1, state.getExpDrop(world, pos1, 0, 0));
             }
         }
